fastapi/requirements.txt
fastapi==0.111.0
gino==1.0.1
python-dotenv==1.0.1
pydantic==2.7.3
alembic==1.13.1
psycopg2==2.9.9